+++ web/src/components/inspector/index.ts
@@ -1,3 +1,4 @@
 export { InspectorHeader } from './InspectorHeader';
 export { InspectorSection } from './InspectorSection';
+export { PaneAnchor } from './PaneAnchor';
 export { InspectorTabList, InspectorTabButton } from './InspectorTabList';